+++ public/js/app.js
@@ -75,7 +75,7 @@ module.exports = __webpack_require__(2);
 /* 1 */
 /***/ (function(module, exports) {
 
-var app = angular.module('items', ["ngRoute"]).constant('API_URL', '/api/');
+var app = angular.module('items', ['ngRoute']).constant('API_URL', '/api/');
 
 app.config(function ($routeProvider) {
     $routeProvider.when('/', {
@@ -107,36 +107,58 @@ app.filter('startFrom', function () {
 });
 
 app.controller('itemsListController', function ($scope, $http, API_URL, $filter) {
-    $http.get(API_URL + "items").then(function (response) {
+    $http.get(API_URL + 'items').then(function (response) {
         $scope.items = response.data;
+
+        $scope.sortType = 'id';
+        $scope.sortReverse = false;
 
         $scope.filterItems = $scope.items;
         $scope.currentPage = 0;
         $scope.itemsPerPage = 5;
         $scope.search = '';
 
+        $scope.toggleSort = function ($event) {
+            $scope.sortType = angular.element($event.currentTarget).attr("data-sort");
+            $scope.sortReverse = !$scope.sortReverse;
+
+            angular.element(document.querySelectorAll('.sort')).removeClass('asc desc');
+
+            angular.element($event.currentTarget).addClass(['asc', 'desc'][+$scope.sortReverse]).removeClass(['asc', 'desc'][+!$scope.sortReverse]);
+
+            $scope.filterItems = $filter('orderBy')($scope.items, $scope.sortType, $scope.sortReverse);
+            console.log($scope.filterItems);
+        };
+
         $scope.firstPage = function () {
             return $scope.currentPage == 0;
         };
+
         $scope.lastPage = function () {
             var lastPageNum = Math.ceil($scope.filterItems.length / $scope.itemsPerPage - 1);
             return $scope.currentPage == lastPageNum;
         };
+
         $scope.getFilterItems = function () {
             return $filter('filter')($scope.filterItems, $scope.search);
         };
+
         $scope.numberOfPages = function () {
             return Math.ceil($scope.getFilterItems().length / $scope.itemsPerPage);
         };
+
         $scope.startingItem = function () {
             return $scope.currentPage * $scope.itemsPerPage;
         };
+
         $scope.pageBack = function () {
             $scope.currentPage = $scope.currentPage - 1;
         };
+
         $scope.pageForward = function () {
             $scope.currentPage = $scope.currentPage + 1;
         };
+
         $scope.$watch('search', function (newValue, oldValue) {
             if (oldValue != newValue) {
                 $scope.currentPage = 0;
@@ -153,9 +175,9 @@ app.controller('itemDetailController', function ($scope, $http, $httpParamSerial
     var url = '';
 
     if (itemId) {
-        url = API_URL + "items/" + itemId;
+        url = API_URL + 'items/' + itemId;
     } else {
-        url = API_URL + "items";
+        url = API_URL + 'items';
     }
 
     $http.get(url).then(function (response) {
@@ -166,13 +188,12 @@ app.controller('itemDetailController', function ($scope, $http, $httpParamSerial
         var fd = new FormData(),
             f = fileInput.files[0];
 
-        fd.append("image", f);
-        fd.append("title", $scope.item.title);
-        fd.append("price", $scope.item.price);
-        fd.append("description", $scope.item.description);
+        fd.append('image', f);
+        fd.append('title', $scope.item.title);
+        fd.append('price', parseFloat($scope.item.price.replace(',', '.')) || 0);
+        fd.append('description', $scope.item.description || '');
 
         $http.post(url, fd, {
-            withCredentials: true,
             headers: { 'Content-Type': undefined },
             transformRequest: angular.identity
         }).then(function (response) {
@@ -182,13 +203,15 @@ app.controller('itemDetailController', function ($scope, $http, $httpParamSerial
     };
 
     $scope.delete = function () {
-        $http({
-            method: 'DELETE',
-            url: url
-        }).then(function (response) {
-            console.log(response);
-            $location.path('/');
-        });
+        if (confirm('Уверены?')) {
+            $http({
+                method: 'DELETE',
+                url: url
+            }).then(function (response) {
+                console.log(response);
+                $location.path('/');
+            });
+        }
     };
 
     $scope.checkDirty = function () {
@@ -199,12 +222,12 @@ app.controller('itemDetailController', function ($scope, $http, $httpParamSerial
             elem.MaterialTextfield.checkDirty();
         });
 
-        document.getElementById("title").setAttribute('required', true);
+        //document.getElementById('title').setAttribute('required', true);
     };
 
-    var fileInputTextDiv = document.getElementById('file_input_text_div');
-    var fileInput = document.getElementById('file_input_file');
-    var fileInputText = document.getElementById('file_input_text');
+    var fileInputTextDiv = document.getElementById('file-input-text-div');
+    var fileInput = document.getElementById('file-input-file');
+    var fileInputText = document.getElementById('file-input-text');
 
     fileInput.addEventListener('change', changeInputText);
     fileInput.addEventListener('change', changeState);
@@ -222,11 +245,11 @@ app.controller('itemDetailController', function ($scope, $http, $httpParamSerial
 
     function changeState() {
         if (fileInputText.value.length != 0) {
-            if (!fileInputTextDiv.classList.contains("is-focused")) {
+            if (!fileInputTextDiv.classList.contains('is-focused')) {
                 fileInputTextDiv.classList.add('is-focused');
             }
         } else {
-            if (fileInputTextDiv.classList.contains("is-focused")) {
+            if (fileInputTextDiv.classList.contains('is-focused')) {
                 fileInputTextDiv.classList.remove('is-focused');
             }
         }
